Garbage collection can run in background task (#502).
Once triggered, a garbage collection run is performed in the
background in a thread. It attempts to lock the cache manager for the
shortest possible time so that lookups can still be done and should
experience only little longer delays while garbage collection is
active.

Garbage collection is potentially triggered at initialization time and
whenever a new stream key is inserted into the cache.

Change-Id: I0e3a50fdfbc3f19a6cae4d08cfc474b99f068d0d